requirements.txt
appdirs==1.4.4
autogluon.common==1.3.0
autogluon.core==1.3.0
autogluon.features==1.3.0
autogluon.tabular==1.3.0
bidict==0.23.1
blinker==1.9.0
boto3==1.38.11
botocore==1.38.11
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dnspython==2.7.0
eventlet==0.39.1
Flask==3.1.0
Flask-SocketIO==5.5.1
fonttools==4.57.0
greenlet==3.2.1
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
kiwisolver==1.4.7
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.9.4
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psutil==7.0.0
pyparsing==3.2.3
pyshark==0.6
python-dateutil==2.9.0.post0
python-engineio==4.12.0
python-socketio==5.13.0
pytz==2025.2
requests==2.32.3
s3transfer==0.12.0
scikit-learn==1.6.1
scipy==1.13.1
simple-websocket==1.1.0
six==1.17.0
termcolor==3.1.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.21.0
